Make numbers be `i64` instead of `usize`

+++ src/evaluation.rs
@@ -1,11 +1,11 @@
 use crate::parsing::{BinaryOperation, BinaryOperator, Expr};
 
 pub trait Eval {
-    fn eval(&self) -> usize;
+    fn eval(&self) -> i64;
 }
 
 impl Eval for Expr {
-    fn eval(&self) -> usize {
+    fn eval(&self) -> i64 {
         match self {
             Expr::Int(_, val) => *val,
             Expr::BinaryOperation(bin_op) => bin_op.eval(),
@@ -15,7 +15,7 @@ impl Eval for Expr {
 }
 
 impl Eval for BinaryOperation {
-    fn eval(&self) -> usize {
+    fn eval(&self) -> i64 {
         match self.operator {
             BinaryOperator::Plus => self.lhs.eval() + self.rhs.eval(),
             BinaryOperator::Times => self.lhs.eval() * self.rhs.eval(),
